% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_pdf.R
\name{analisis.pdf.latex}
\alias{analisis.pdf.latex}
\title{Informe Descriptivo en formato PDF y Latex}
\usage{
analisis.pdf.latex(base, colhist, colbox, name, direc)
}
\arguments{
\item{base}{Nombre base de datos a analizar, esta debe contener variables de tipo
continuas.}

\item{colhist}{Color del Histograma a generar, debe ser introducido entre comillas.
Ej: "green".}

\item{colbox}{Color de la grafica Box-Plot a generar, debe ser introducido entre comillas.
Ej: "blue".}

\item{name}{Corresponde al nombre del archivo PDF generado, este debe introducirse
de la siguiente forma: "nombre.pdf"}

\item{direc}{Ubicación donde se guaradara el archivo generado, debe ser introducido
entre comillas. Ej:"C:/Users/usuario/Desktop"}
}
\description{
Genera un informe estadistico que incorpora estadisticos descritivos
   , correlaciones entre variables, histogramas de las variables y graficas Box Plot.
}
