% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_de_variacion.R
\name{coeficiente.variacion}
\alias{coeficiente.variacion}
\title{Coeficiente de Variacion}
\usage{
coeficiente.variacion(x)
}
\arguments{
\item{x}{Nombre de la variable.}
}
\description{
Calcula el Coeficiente de Variacion de una variable de tipo continua,
este relaciona la media y la variabilidad de la variable.
}
\examples{
data("albahaca")
coeficiente.variacion(albahaca$produccion)
}
\references{
Casella, G. (1990). Statistical Inference. Duxburry Press.
}
