% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matriz_pearson.R
\name{corr.matrix.pearson}
\alias{corr.matrix.pearson}
\title{Matriz de correlacion de Pearson}
\usage{
corr.matrix.pearson(data.base)
}
\arguments{
\item{data.base}{Nombre de la base de datos}
}
\description{
Calcula la matriz de correlacion de Pearson para las variables continuas
    presentes en una base de datos.
}
\examples{
data("albahaca")
corr.matrix.pearson(albahaca)
}
\references{
Casella, G. (1990). Statistical Inference. Duxburry Press.
}
