% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix.R
\name{corr.matrix}
\alias{corr.matrix}
\title{Matriz de correlacion de Pearson}
\usage{
corr.matrix(base, cont)
}
\arguments{
\item{base}{Nombre de la base de datos}

\item{cont}{Vector nombre de variables continuas.}
}
\description{
Calcula la matriz de correlaciones de Pearson para las variables continuas
    de una base de datos.
}
\examples{
data("albahaca")
corr.matrix(albahaca,c("produccion","temp","alt","n.riegos"))
}
\references{
Casella, G. (1990). Statistical Inference. Duxburry Press.
}
