% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matriz_spearman.R
\name{corr.matrix.spearman}
\alias{corr.matrix.spearman}
\title{Matriz de correlacion de Spearman}
\usage{
corr.matrix.spearman(data.base)
}
\arguments{
\item{data.base}{Nombre de la base de datos}
}
\description{
Calcula la matriz de correlacion de Spearman para las variables continuas
    presentes en una base de datos.
}
\examples{
data("albahaca")
corr.matrix.spearman(albahaca)
}
\references{
Casella, G. (1990). Statistical Inference. Duxburry Press.
}
