% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_pearson.R
\name{corr.pearson}
\alias{corr.pearson}
\title{Correlacion de Pearson}
\usage{
corr.pearson(base, x, y)
}
\arguments{
\item{base}{Nombre de la base de datos.}

\item{x}{Nombre variable X.}

\item{y}{Nombre variable Y.}
}
\description{
Calcula la correlacion entre dos variables de tipo continuas junto
    con su respectiva interpretacion, ademas entrega una grafica de dispercion y
    tendencia lineal.
}
\references{
Casella, G. (1990). Statistical Inference. Duxburry Press.
}
