% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_spearman.R
\name{corr.spearman}
\alias{corr.spearman}
\title{Matriz de correlacion de Spearman}
\usage{
corr.spearman(data.base)
}
\arguments{
\item{data.base}{Nombre de la base de datos}
}
\description{
Calcula la matriz de correlaciones de Spearman para las variables continuas
   de una base de datos.
}
\examples{
data("albahaca")
corr.spearman(albahaca)
}
\references{
Casella, G. (1990). Statistical Inference. Duxburry Press.
}
