% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianza.R
\name{covarianza}
\alias{covarianza}
\title{Covarianza}
\usage{
covarianza(x, y)
}
\arguments{
\item{x}{Nombre variable X.}

\item{y}{Nombre variable Y.}
}
\description{
Calcula la covarianza entre dos variables de tipo continuas.
}
\examples{
data("albahaca")
covarianza(albahaca$produccion,albahaca$temp)
}
\references{
Casella, G. (1990). Statistical Inference. Duxburry Press.
}
