% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivos_categoricas.R
\name{descriptive.categorical}
\alias{descriptive.categorical}
\title{Estadisticos Descriptivos Variables Categoricas no ordinales}
\usage{
descriptive.categorical(data.base)
}
\arguments{
\item{data.base}{Nombre de la base de datos.}
}
\description{
Entrega un resumen descriptivo de las variables categoricas de una base de datos
   que contine variables de tipo numericas y categoricas no ordinales.
}
\examples{
data("albahaca")
descriptive.categorical(albahaca)
}
\references{
Casella, G. (1990). Statistical Inference. Duxburry Press.
}
