% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivos.R
\name{descriptive.continue}
\alias{descriptive.continue}
\title{Estadisticos Descriptivos}
\usage{
descriptive.continue(base)
}
\arguments{
\item{base}{Nombre de la base de datos.}
}
\description{
Estrega un resumen descriptivo de una base de datos que posee variables
   continuas. Permite la obtención del Numero de datso, Promedio, Suma, Desviacion Estandar,
   Coeficiente de Variacion, Asimetria, Curtosis, Minimo y Maximo.
}
\examples{
data("albahaca")
descriptive.continue(albahaca)
}
\references{
Casella, G. (1990). Statistical Inference. Duxburry Press.
}
