% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_regresion.R
\name{descriptive.regression}
\alias{descriptive.regression}
\title{Correlacion de Pearson y Coeficientes de Regresion Lineal Simple}
\usage{
descriptive.regression(data.base)
}
\arguments{
\item{data.base}{Nombre de la base de datos.}
}
\description{
Calcula la correlacion de Pearson entre las combinaciones posibles de
   variables continuas de una base de datos, entrega una estimacion de los coeficientes de Regresion Lineal.
}
\examples{
data("albahaca")
cuartiles(albahaca$produccion)
}
\references{
Casella, G. (1990). Statistical Inference. Duxburry Press.
}
