% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_pdf.R
\name{informe.pdf}
\alias{informe.pdf}
\title{Informe Descriptivo}
\usage{
informe.pdf(base, name, direc)
}
\arguments{
\item{base}{Nombre base de datos a analizar, esta debe contener variables de tipo
continuas.}

\item{name}{Nombre archivo a generar.}

\item{direc}{Diarectorio donde se almacena el archivo.}
}
\description{
Genera un informe estadistico que incorpora estadisticos descritivos
   , correlaciones entre variables, histogramas de las variables y graficas Box Plot.
}
\examples{
data("albahaca")
informe.word(albahaca,"estadisticos.pdf","C:/Users/usuario/Desktop")
}
