% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_word.R
\name{informe.word}
\alias{informe.word}
\title{Informe Estadisticos Descriptivos formato Word}
\usage{
informe.word(base, name, direc)
}
\arguments{
\item{base}{Nombre base de datos a analizar.}

\item{name}{Nombre archivo a generar,este argumento debe ser entregado en formato:
"nombre.docx"}

\item{direc}{Directorio donde se almacena el archivo a generar.}
}
\description{
Genera un informe estadistico que incorpora estadisticos descritivos
   , correlaciones entre variables, Histogramas y graficas Box Plot.
}
\examples{
\dontrun{
data("albahaca")
informe.word(albahaca,"estadisticos.docx","C:/Users/usuario/Desktop")}
}
