% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximo.R
\name{maximo}
\alias{maximo}
\title{Maximo}
\usage{
maximo(x)
}
\arguments{
\item{x}{Nombre de la variable.}
}
\description{
Encuenta el maximo valor de una variable continua.
}
\examples{
data("albahaca")
maximo(albahaca$produccion)
}
