% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordena.R
\name{ordena}
\alias{ordena}
\title{Ordena}
\usage{
ordena(x)
}
\arguments{
\item{x}{Nombre de la variable.}
}
\description{
Ordena de menor a mayor una variable de tipo continua.
}
\examples{
data("albahaca")
ordena(albahaca$produccion)
}
