% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media.R
\name{promedio}
\alias{promedio}
\title{Promedio}
\usage{
promedio(x)
}
\arguments{
\item{x}{Nombre de la variable.}
}
\description{
Encuenta el promedio de una variable de tipo continua.
}
\examples{
data("albahaca")
promedio(albahaca$produccion)
}
\references{
Casella, G. (1990). Statistical Inference. Duxburry Press.
}
