% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernulli.R
\name{sim.bernulli}
\alias{sim.bernulli}
\title{Simulacion Variable Aleatoria Bernulli}
\usage{
sim.bernulli(n, p)
}
\arguments{
\item{n}{Numero de valores aleatorios a generar.}

\item{p}{Probabilidad de exito.}
}
\description{
Genera valores aleatorios provenienes de una distribucion Bernulli.
}
\examples{
sim.bernulli(100,0.4)
}
\references{
Ross, S.M. (1999) Simulacion. Prentice Hall.
}
