% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{sim.binomial}
\alias{sim.binomial}
\title{Simulacion Variable Aleatoria Binomial}
\usage{
sim.binomial(m, n, p)
}
\arguments{
\item{m}{Numero de valores a generar.}

\item{n}{Numero de ensayos.}

\item{p}{Probabilidad de exito.}
}
\description{
Genera valores aleatorios provenienes de una distribucion Binomial.
}
\examples{
sim.binomial(100,30,0.4)
}
\references{
Ross, S.M. (1999) Simulacion. Prentice Hall.
}
