% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponencial.R
\name{sim.exponencial}
\alias{sim.exponencial}
\title{Simulacion Variable Aleatoria Exponencial}
\usage{
sim.exponencial(n, L)
}
\arguments{
\item{n}{Numero de valores aleatorios a generar.}

\item{L}{Parametro Lambda.}
}
\description{
Genera valores aleatorios provenienes de una distribucion Exponencial
   junto con un histograma en el cual se comparan los valores simulados con la densidad
   teorica de la variable aleatoria simulada.
}
\examples{
sim.exponencial(100,7)
}
\references{
Ross, S.M. (1999) Simulacion. Prentice Hall.
}
