% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometrica.R
\name{sim.geometrica}
\alias{sim.geometrica}
\title{Simulacion Variable Aleatoria Geometrica}
\usage{
sim.geometrica(n, p)
}
\arguments{
\item{n}{Numero de valores aleatorioas a generar.}

\item{p}{Probabilidad de exito.}
}
\description{
Genera valores aleatorios provenienes de una distribucion Geometrica.
}
\examples{
sim.geometrica(100,0.4)
}
\references{
Ross, S.M. (1999) Simulacion. Prentice Hall.
}
