% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hipergeometrica.R
\name{sim.hipergeometrica}
\alias{sim.hipergeometrica}
\title{Simulacion Variables Aleatorias Hipergeometricas}
\usage{
sim.hipergeometrica(t, N, N1, n)
}
\arguments{
\item{t}{Numero de valores aleatorios a generar.}

\item{N}{Numero de elementos en la poblacion.}

\item{N1}{Numero de elementos en la poblacion que pertenecen a la categoria deseada.}

\item{n}{Tamaño muestral.}
}
\description{
Genera valores aleatorios provenienes de una distribucion Hipergeometrica
   junto con un histograma en el cual se comparan los valores simulados con la densidad
   teorica de la variable aleatoria simulada.
}
\examples{
sim.hipergeometrica(200,100,40,20)
}
\references{
Ross, S.M. (1999) Simulacion. Prentice Hall.
}
