% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisson.R
\name{sim.poisson}
\alias{sim.poisson}
\title{Simulacion Variable Aleatoria Poisson}
\usage{
sim.poisson(n, L)
}
\arguments{
\item{n}{Numero de valores aleatorios a generar.}

\item{L}{Parametro Lambda, el cual representa el numero de veces que se espera que ocurra el fenomeno
durante un intervalo de tiempo.}
}
\description{
Genera valores aleatorios provenienes de una distribucion Poisson
   junto con un histograma en el cual se comparan los valores simulados con la densidad
   teorica de la variable aleatoria simulada.
}
\examples{
sim.poisson(100,5)
}
\references{
Ross, S.M. (1999) Simulacion. Prentice Hall.
}
