% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_bivariada_box_muller.R
\name{snormalBM}
\alias{snormalBM}
\title{Simulacion Variable Aleatoria Normal Bivariada via Box Muller}
\usage{
snormalBM(n)
}
\arguments{
\item{n}{Numero de valores aleatorios a generar.}
}
\description{
Genera valores aleatorios simulados que siguen una distribucion Normal con
   media cero y varianza uno aplicando el metodo de Box-Muller.
}
\examples{
snormalBM(100)
snormalBM(200)
}
\references{
Ross, S.M. (1999) Simulacion. Prentice Hall
}
