% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_bivariada_coordenadas_polares.R
\name{snormalCP}
\alias{snormalCP}
\title{Simulacion Variable Aleatoria Normal Bivariada via Coordenadas Polares}
\usage{
snormalCP(n)
}
\arguments{
\item{n}{Numero de valores aleatorios a generar.}
}
\description{
Genera valores aleatorios simulados que siguen una distribucion Normal con
   media cero y varianza uno, estos valores se obtienen aplicando el metodo de Coordenadas Polares.
}
\examples{
snormalCP(100)
snormalCP(50)
}
\references{
Ross, S.M. (1999) Simulacion. Prentice Hall.
}
