% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std.R
\name{std}
\alias{std}
\title{Desviacion Estandar}
\usage{
std(x)
}
\arguments{
\item{x}{Nombre de la variable.}
}
\description{
Calcula la Desviacion Estandar de una variable de tipo continua.
}
\references{
Casella, G. (1990). Statistical Inference. Duxburry Press.
}
