% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma}
\usage{
suma(x)
}
\arguments{
\item{x}{Nombre de la variable.}
}
\description{
Calcula la suma de una variable continua.
}
\examples{
data("albahaca")
suma(albahaca$produccion)
}
