% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix.R
\name{corr.matrix}
\alias{corr.matrix}
\title{Matrix de correlaciones}
\usage{
corr.matrix(base, cont)
}
\arguments{
\item{base}{Nombre de la base de datos}

\item{cont}{Vector nombre de variables continuas.}
}
\description{
Calcula la matriz de correlaciones de una base de una base de datos
    que contiene variables de tipo continuas.
}
\references{
Casella, G. (1990). Statistical Inference. Duxburry Press.
}
