% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_pearson.R
\name{corr.pearson}
\alias{corr.pearson}
\title{Correlacion de Pearson}
\usage{
corr.pearson(base, x, y)
}
\arguments{
\item{base}{Nombre de la base de datos.}

\item{x}{Nombre variable X.}

\item{y}{Nombre variable Y.}
}
\description{
Calcula la correlacion de Pearson entre dos variables de tipo continuas,
   entregando una interpretacion de los resultados obtenidos y entregando una grafica
   que presenta la dispercion entre las variables y la recta que se obtendria al aplicar
   un metodo de Regresion Lineal Simple.
}
\examples{
data("albahaca")
corr.pearson(albahaca,"n.riegos","produccion")
}
\references{
Casella, G. (1990). Statistical Inference. Duxburry Press.
}
