% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivos.R
\name{descriptive.continue}
\alias{descriptive.continue}
\title{Estadisticos Descriptivos Variables Continuas}
\usage{
descriptive.continue(base)
}
\arguments{
\item{base}{Nombre de la base de datos.}
}
\description{
Entrega un resumen descriptivo de las variables continuas de una base de datos.
   Permite la obtención del Numero de datos, Promedio, Suma, Desviacion Estandar,
   Coeficiente de Variacion, Asimetria, Curtosis, Minimo, Maximo y Cuartiles.
}
\examples{
data("albahaca")
descriptive.continue(albahaca)
}
\references{
Casella, G. (1990). Statistical Inference. Duxburry Press.
}
