% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivos.R
\name{descriptive}
\alias{descriptive}
\title{Estadisticos Descriptivos}
\usage{
descriptive(base, cont)
}
\arguments{
\item{base}{Nombre de la base de datos.}

\item{cont}{Vector que contine el nombre de las variables continuas a analizar.}
}
\description{
Estrega un resumen descriptivo de una base de datos que posee variables
   continuas. Permite la obtención del Numero de datso, Promedio, Suma, Desviacion Estandar,
   Coeficiente de Variacion, Asimetria, Curtosis, Minimo y Maximo.
}
\references{
Casella, G. (1990). Statistical Inference. Duxburry Press.
}
