% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_pdf.R
\name{informe.pdf}
\alias{informe.pdf}
\title{Informe Estadisticos Descriptivos formato PDF}
\usage{
informe.pdf(base, name, direc)
}
\arguments{
\item{base}{Nombre de base de datos a analizar.}

\item{name}{Nombre archivo a generar, este argumento debe ser entregado en formato:
"nombre.pdf"}

\item{direc}{Directorio donde se almacena el archivo a generar.}
}
\description{
Genera un informe estadistico que incorpora estadisticos descritivos
   , correlaciones entre variables, Histogramas y graficas Box Plot.
}
\examples{
\dontrun{
data("albahaca")
informe.pdf(albahaca,"estadisticos.pdf","C:/Users/usuario/Desktop")}
}
