% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_word.R
\name{informe.word}
\alias{informe.word}
\title{Informe Descriptivo Word}
\usage{
informe.word(base, cont, cat, name, direc)
}
\arguments{
\item{base}{Nombre base de datos a analizar, esta debe contener variables de tipo
continuas.}

\item{cont}{Vector nombre variables continuas.}

\item{cat}{Vector nombre variables categoricas.}

\item{name}{Nombre archivo a generar.}

\item{direc}{Diarectorio donde se almacena el archivo.}
}
\description{
Genera un informe estadistico que incorpora estadisticos descritivos
   , correlaciones entre variables, histogramas de las variables y graficas Box Plot.
}
